pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "piassistant"
version = "0.1.0"
description = "Smart assistant on Raspberry Pi 5 with Claude AI brain"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.100",
    "uvicorn[standard]>=0.24",
    "anthropic>=0.42",
    "httpx>=0.25",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
piassistant = "piassistant.main:run_server"
piassistant-cli = "piassistant.main:run_cli"

[tool.pytest.ini_options]
asyncio_mode = "strict"
